[Issue #2493] Allow certain agency fields to be null from the legacy values (#2513)
## Summary
Fixes #2493

### Time to review: __5 mins__

## Changes proposed
Modified three fields (ldapGp, description, label) in the agency table
to be nullable

Adjusted the transform agency process to allow these fields to be
nullable as well

## Context for reviewers
This isn't an issue in prod right now, but is accounting for a problem I
found in the dev/staging environments. The agency data is stored in a
very different way (key-value pairs) in the legacy data, and these
fields are sometimes not present in the test data. There isn't any
reason we should require them as we likely won't use them, so I'm
letting them be nullable to let these agencies with the values null get
populated in the lower envs.

## Additional information
There is some additional context on
https://github.com/HHS/simpler-grants-gov/issues/2494

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>