[Issue #2389] Test and adjust Opportunity ORM relationships (#2550)
## Summary
Fixes #2389 

### Time to review: __5 mins__

## Changes proposed
Add tests to validate the behavior of our opportunity relationships,
mostly in regards to deletion events

Fix one faulty relationship

Fix a mypy issue regarding the interactive console script (a recent mypy
upgrade doesn't like this script's imports, so move it to a place that
won't get checked)

## Context for reviewers
Our ORM tries to be smart / we can configure our ORM to be smart when we
delete an object. For example, if we do a
`db_session.delete(some_opportunity_record)` the ORM will first figure
out what depends on this record via our relationships, and delete other
records (for example, an opportunity summary) in the proper order saving
us a lot of complexity in the deletes.

However, sometimes the ORM doesn't quite choose right (deleting a child
object deletes the whole thing). There have been a few bug fixes I've
made to address this, and these tests confirm we've got those issues
worked out + found another one.

Without the change I made here, deleting an opportunity makes the ORM
try to null out a non-null column in the agency table (why that is the
default behavior, I'm not sure). By making it a `viewonly` relationship,
the ORM knows not to cascade across the relationships.