[Issue #2472] Remove all ECS terraform stuff (#2555)
## Summary

Relates to https://github.com/HHS/simpler-grants-gov/issues/2472, undos
https://github.com/HHS/simpler-grants-gov/pull/2480

### Time to review: __0.1 mins__

## Changes proposed

Removes all the ecs-terraform stuff

## Context for reviewers

I spent more time on this than my (self imposed) timebox, so I'm
swapping to the VPN route instead. If nothing else, it was in fact fun
to work on. What is engineering if not the PRs we made along the way.