[Issue #2548] Swap analytics to new step function pattern (#2546)
## Summary

Relates to #2548

### Time to review: __2 mins__

## Changes proposed

Swaps the analytics application to the new step functions pattern that I
originally merged in
https://github.com/HHS/simpler-grants-gov/pull/2506. The size of the red
diff and the simplicity of the new configuration show have valuable the
new pattern is.

## Context for reviewers

The step function is now running, but the CLI itself is currently
broken. eg. the CLI is currently broken everywhere, even locally. Fixing
that is someone else's job though!