[Issue #2528] Change job schedule ELT process to hourly (#2641)
## Summary
Fixes #2528

### Time to review: __3 mins__

## Changes proposed
Modified ELT process to run hourly

Removed the load job for the v0 data

## Context for reviewers
I can't imagine there is much value in the v0 job now that v1 works.
Probably will save us a bit as we won't be running a process near
constantly that no one uses.

I changed the schedule expression to a cron so that the schedule is a
bit more consistent. The `rate(..)` configuration seems to be based on
when you turn it on, or last did a deploy. This will keep it from
varying every time we deploy and instead make it so it's always at the
top of the hour.

## Additional information

https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-scheduled-rule-pattern.html