[Issue #2630] Enable search load to run hourly (#2689)
## Summary
Fixes #2630

### Time to review: __3 mins__

## Changes proposed
Schedule the DB -> search index load job to run every hour

Set the shard/replica count for prod as it needs to be higher as we have
more nodes.

## Context for reviewers
In the future we may want to coordinate the ELT process with this (put
them in the same flow?), but for now I've setup the job to run 30
minutes after the start of the hour. The job in prod takes ~9 minutes at
the moment.

The ELT process takes 4-5 minutes and runs at the top of the hour - the
jobs overlapping wouldn't be a big deal other than search having staler
data.