[Issue #2697] Enable agency data load in ELT process (#2698)
## Summary
Fixes #2697

### Time to review: __5 mins__

## Changes proposed
Enable the extract-load process on the TGROUPS table (agency data)

Enable the Agency transform steps

## Context for reviewers
I've tested this in the lower environments and prod by manually
configuring the job to load the data, so all of the agency data is there
already. This just gets updates from here on out.

The extract-load process has a set of tables it is configured to load,
and the transformation process has specific classes for transforming
tables that are handled via env vars, so the configuration change is
minor.

The unit test changes I made show that the agency transforms are
happening now when running the "full" job.