[no ticket] Fix infinite state locks (#2779)
### Time to review: __1 mins__

## Context for reviewers

Platform's assertion is this: whenever a deploy fails for any reason, it
cancels the deploy, which locks the other 3 jobs. Those 3 jobs remain
locked indefinitely. On the next deploy, every job but 1 is locked, but
the other 3 jobs fail because they were locked prior, which causes 1
first job to be canceled, and thusly all 4 jobs are locked. It's an
avalanche effect. Whenever 1 deploy fails, all 4 fail that point
onwards.