[Issue 2665] Analytics jobs conditionally use IAM token as Postgres pwd (#2799)
## Summary
Fixes #2665 

### Time to review: __2 mins__

## Changes proposed
> What was added, updated, or removed in this PR.

Adds a switch in `analytics/integrations/db.py` that determines which
value to use for a DB password when connecting to Postgres: either the
value in `local.env` or an IAM token, depending on an environment
variable.

Also adds better exception handling, to make errors easier to spot in
CI.

## Context for reviewers
> Testing instructions, background context, more in-depth details of the
implementation, and anything else you'd like to call out or ask
reviewers. Explain how the changes were verified.

This is a follow up to previous PRs,
https://github.com/HHS/simpler-grants-gov/pull/2786 and
https://github.com/HHS/simpler-grants-gov/pull/2796, and part of an
effort to get `analytics` step functions to successfully connect to
Postgres DB in a CI environment.

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.