[Issue #2721] Setup logic to validate login.gov tokens against their public keys (#2837)
## Summary
Fixes #2721

### Time to review: __15 mins__

## Changes proposed
Add logic to validate the authenticity of tokens from login.gov

Add rough logic for fetching public keys from login.gov

Add logic for parsing tokens from login.gov, extracting relevant info

## Context for reviewers
We will be processing the id_token returned by this login.gov endpoint
https://developers.login.gov/oidc/token/

This token follows most general JWT conventions. In order to validate
the authenticity of the token, we need to fetch the public key from
login.gov (see: https://developers.login.gov/oidc/certificates/ ). This
is all standard Oauth stuff, we just have to deal with complexities
surrounding the caching of these public keys.

We assume that if a token fails to validate with the public key, that it
might be with a key we don't yet know about, and refresh the keys (this
is a pretty standard approach), however this refresh runs into
thread-safety issues that are documented inline. In short, the API runs
multiple threads and they might all try to refresh simultaneously.
Whether that is an issue, we'll have to see.

Most of the actual validation on the token is done inside of the library
and just boils down to:
* Proving it was generated with the key (the public key verification the
JWT library handles
* Verifying timestamps are in the right order
* Verifying who sent us this token (clientID will need to be configured
to match whatever the frontend sets up).

After all this, we're certain we know who they are, and can use the
token, although the current logic stops after parsing.

## Additional information
More thorough testing of this behavior needs the frontend / integration
with login.gov to be a bit further along. There are a few small unknowns
in the details of the token we'll be parsing that might require small
adjustments.

To avoid committing tokens, I make the tests fully generate the RSA
tokens in the tests. This is the same as defining a public and private
key in a very long form (the "BEGIN PUBLIC KEY..." files).