[Issue #2882] Setup logging in analytics codebase (#2884)
## Summary
Fixes #2882

### Time to review: __30 mins__

## Changes proposed
Logging - a lot of it - based on what the API already does with some
stuff trimmed out

Highlights:
* Python logging configured and usable with minimal effort
* Locally logging defaults to a human-readable format with some
coloring, non-locally written out as JSON
* PII (SSN/TIN) auto-masking in all parts of a log
* Formatting of logs is ready for easy querying in Cloudwatch and New
Relic
* Automatically append basic information to all log messages (eg.
environment, task name)
* Automatically fetch basic ECS task information and attach to log
messages

## Context for reviewers
This is ~90% copy-paste from the API code base with two notable
exceptions:
* Linting rules are much stricter in analytics, so adjustments to
formatting were made. Mostly docstring related
* The API has additional logic for things like request/response logging
and other API and/or Flask specific concepts.

Assuming the new `init` method added to the cli file is always called
(as we add more commands, probably want to split that into a few files),
then logging will work anywhere. You simply need to do:

```
import logging

logger = logging.getLogger(__name__)
```
at the top of your file and that logger will _just work_. You can add
extra values in the form of a dict that will get added to the overall
JSON output for a log message.

Note that in order to get the ECS task stuff a decorator needs to be
added to the entrypoint method of your script which I just added to the
initialize_database command for the moment.

## Additional information
An example of the JSON logs generated by the task automatically (in this
case, the last message saying it was successful):
```json
{
  "name": "analytics.logs.ecs_background_task",
  "msg": "Completed ECS task %s",
  "args": [
    "initialize_database"
  ],
  "levelname": "INFO",
  "levelno": 20,
  "pathname": "/analytics/src/analytics/logs/ecs_background_task.py",
  "filename": "ecs_background_task.py",
  "module": "ecs_background_task",
  "exc_info": null,
  "exc_text": null,
  "stack_info": null,
  "lineno": 82,
  "funcName": "_ecs_background_task_impl",
  "created": 1731706528.165175,
  "msecs": 165,
  "relativeCreated": 4169.229745864868,
  "thread": 281473801429024,
  "threadName": "MainThread",
  "processName": "MainProcess",
  "process": 1,
  "taskName": null,
  "ecs_task_duration_sec": 0.148,
  "status": "success",
  "environment": null,
  "task_name": "initialize_database",
  "task_uuid": "136089d8-b765-459d-9f59-d7f8cd9c8323",
  "message": "Completed ECS task initialize_database"
}
```

An example of logs when there is an exception:
```json
{
  "name": "analytics.logs.ecs_background_task",
  "msg": "ECS task failed",
  "args": [],
  "levelname": "ERROR",
  "levelno": 40,
  "pathname": "/analytics/src/analytics/logs/ecs_background_task.py",
  "filename": "ecs_background_task.py",
  "module": "ecs_background_task",
  "exc_info": [
    "<class 'Exception'>",
    "An example error",
    "<traceback object at 0xffff8ab4ffc0>"
  ],
  "exc_text": "Traceback (most recent call last):\n  File \"/analytics/src/analytics/logs/ecs_background_task.py\", line 72, in _ecs_background_task_impl\n    yield\n  File \"/analytics/src/analytics/logs/ecs_background_task.py\", line 52, in wrapper\n    return f(*args, **kwargs)\n           ^^^^^^^^^^^^^^^^^^\n  File \"/analytics/src/analytics/cli.py\", line 275, in initialize_database\n    raise Exception(\"An example error\")\nException: An example error",
  "stack_info": null,
  "lineno": 77,
  "funcName": "_ecs_background_task_impl",
  "created": 1731706654.6125245,
  "msecs": 612,
  "relativeCreated": 4152.841091156006,
  "thread": 281473382309920,
  "threadName": "MainThread",
  "processName": "MainProcess",
  "process": 1,
  "taskName": null,
  "status": "error",
  "environment": null,
  "task_name": "initialize_database",
  "task_uuid": "ec23ba72-0c3a-4479-a075-5884518ebfe9",
  "message": "ECS task failed"
}
```

Here are a few screenshots of running the task with human-readable logs
enabled (the default locally). Including one where an exception is
raised.

<img width="734" alt="Screenshot 2024-11-15 at 4 20 21 PM"
src="https://github.com/user-attachments/assets/34740713-3d90-47c7-a50c-1ca1a570c06d">

![Screenshot 2024-11-15 at 4 20
32 PM](https://github.com/user-attachments/assets/a722e696-eee3-46e3-affc-da9c0145d334)

![Screenshot 2024-11-15 at 4 38
51 PM](https://github.com/user-attachments/assets/45d52cb5-1d66-45e5-ab4c-4faa25e86d0d)

For context, here is what the output looked like before these changes:
```
initializing database
done
```