[Issue 2869] Add versioning to analytics db schema (#2870)
## Summary
Fixes #2869 

### Time to review: __5 mins__

## Changes proposed
> What was added, updated, or removed in this PR.

1. Added a simple versioning pattern to enable schema extensibility,
including a table in the schema to keep track of current schema version
2. Refactored `make init-db` command to enable execution of multiple SQL
files, and update schema version accordingly upon the execution of each
file

## Context for reviewers
> Testing instructions, background context, more in-depth details of the
implementation, and anything else you'd like to call out or ask
reviewers. Explain how the changes were verified.

This is part of an effort to enable automated deployment of analytics db
to staging and prod.

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.