[Issue #1884] make artillery very aggressive (#2940)
## Summary

Relates to #1884

### Time to review: __2 mins__

## Changes proposed

Scales the artillery config up to "the site falls down" numbers, so that
we can see exactly what the failure cases are for our website.

Unless we are testing for DDOS protection, we should never need to test
artillery with higher numbers than this. These numbers are _**very**_
high.

## Context for reviewers

We tested this already in staging, and it does in fact break the site.
Or at least, it does with 4 people running it.