[Issue #3058] Static List of Agencies for Search filter was incomplete (#3150)
## Summary
Fixes #3058

### Time to review: __1 mins__

## Changes proposed
Pasting in an [updated JSON
extract](https://github.com/HHS/simpler-grants-gov/issues/3113#issuecomment-2520931188)
that is complete for all Agencies listing Opportunities currently. This
will move to be a live API call in a follow-up ticket.

## Context for reviewers
Testing by validating that the filter list for Agencies is much more
extensive now. Because we're using a static list of Prod Agencies we
can't actually test that the filters work and return Opportunities