[Issue #3102] Call Oauth token endpoint (#3122)
## Summary
Fixes #3102

### Time to review: __10 mins__

## Changes proposed
Added a client (and mock) for calling an OAuth token endpoint

A lot of restructuring of test utils for clearer setup

## Context for reviewers
https://developers.login.gov/oidc/token/

This gets the token from the OAuth server and parses the response. There
is more work to do on this later as login.gov requires a special JWT to
also be passed, but a basic version doesn't need that (our local mock
doesn't care), so I'll follow-up on that later.

This approach to setting up a client is following some patterns I've
used before that worked well. Building a mock version alongside the real
one helps with testing.

## Additional information
Still nothing new visually, under the hood it is just one more big step
remaining to process the token

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>