[Issue #3103] Pass client assertion jwt to token endpoint (#3190)
## Summary
Fixes #3103

### Time to review: __10 mins__

## Changes proposed
Changed the way we call the OAuth token endpoint to use private_key_jwt

Added docs for how to setup Login.gov certs

## Context for reviewers
For the first pass, I setup the call to the token endpoint to use
client_id, but actual Login.gov uses private_key_jwt for this instead
(which includes the client ID). Luckily it seems our local mock is fine
with that and can take in this as well. It doesn't do any validation on
the key itself, so any private key is fine.

Effectively, this form of "auth" is just passing the client ID + a hash
that could have only been created by our private key, and login.gov is
configured to have our public key.

## Additional information
For testing this, I actually connected it to our dev login.gov app which
I setup alongside it. It works! Also verified exactly what we need to
set for many of the env vars to get everything happy.

You get directed first to:

![Screenshot 2024-12-11 at 4 14
39 PM](https://github.com/user-attachments/assets/25474a57-afb8-42dd-8c74-599f4ac3edab)

And then back to our API and then the final endpoint with a token that
works locally.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>