[Issue #3183] Make search facets for agencies return more than 25 agencies (#3273)
## Summary
Fixes #3183

### Time to review: __3 mins__

## Changes proposed
Modify the search facet count logic to allow up to 1000 unique agencies
in response

## Context for reviewers
By default we configured the size to be 25 arbitrarily. There are about
700 agencies with opportunities in prod, so configured a number a bit
above that for now.

We might want to consider allowing the user of the API to specify the
facets they want in the future.

## Additional information
I loaded a lot of opportunities into the search index, and the response
for facet counts now has all of the agencies we have configured locally
with is about 80.