[Issue #3193] Use the KID value in login gov token validation (#3283)
## Summary
Fixes #3193 

### Time to review: __5 mins__

## Changes proposed
Rather than iterate over all public keys from login.gov - instead
directly use the one with the same KID (key ID)

## Context for reviewers
The KID is the Key ID, in the public endpoint from an OAuth server that
gives you a list of keys the ID should be present there, and the header
for a given token will have the KID as well. However these weren't
guaranteed in the OAuth spec so I waited to assume they existed until I
tested with real login.gov - they do exist.

Quite a bit of structure needed to be reorganized with this as the way
we also structure the public keys went from a list to a dict. There was
a lot of assumed failure cases if no key was found that worked that were
removed and now the logic is a bit more directly:
* If a public key can't be found refresh the public keys
* If a public key is found, validate with it
* Otherwise error

We also no longer need to account for the "failed validation, but assume
we used the wrong key" scenario since we'll have the exact key.

## Additional information
Everything continues to work uneventfully locally, the KID is set
(always `issuer1`) in both places.