[Issue #2841] Make the robots.txt dynamic so that it can mutate per environment (#3379)
## Summary
Fixes #2841

### Time to review: __5 mins__

## Changes proposed
We need to make the robots.txt dynamic per environment so that we can
ban crawling of the lower environments but still support nuanced rules
for upper environments.

## Context for reviewers
As a follow up change we will also make the Sitemap dynamic as well
since we don't want to point to Sitemaps outside of Production. Our
Sitemap isn't currently ready for being utilized so not including it for
now.

## Additional information
Tested this locally and confirmed that we do have a dynamic robots.txt
that reacts to changes to the environment variable at runtime, not build
time. Not sure if this will all hook together as expected when it tries
to run in AWS, so having this test for "dev" temporarily until we prove
it is all likely to work in Prod.