[Issue #3271] Setup structure of opportunity attachment transformation (minus file logic) (#3443)
## Summary
Fixes #3271 

### Time to review: __10 mins__

## Changes proposed
Add a subtask to transform the opportunity attachment table

DOES NOT handle the files themselves (next PR will handle that)

## Context for reviewers
The transformation code is bulky and has a lot of repeated patterns that
needed to be setup. This PR gets all of that setup and working for the
opportunity attachment table, except for the logic regarding moving the
file from the staging table to our s3 bucket (which has a lot of extra
fun complexities that will need to be handled special). Every other
column in the table we want is copied over and functionally working with
these changes.

The follow-up to this likely will refactor a bit to make the file logic
reasonable, but figured it was best to start with a pattern that looks
like the other transformations we've done.

## Additional information
From a clean database, you can run this locally by doing `make console`
and then using our factories to generate data
```py
f.StagingTsynopsisAttachmentFactory.create_batch(size=50)
exit()
```

I then modified `TransformOracleDataTask` to only run the
TransformOpportunity and new TransformOpportunityAttachment classes
which successfully copied to my local opportunity_attachment table as
expected:
![Screenshot 2025-01-07 at 3 28
25 PM](https://github.com/user-attachments/assets/f7183af3-97fa-4de1-a963-3a6eefea7bb9)