[Unticketed] Fix issue where opportunities aren't loaded to search index (#3567)
## Summary

### Time to review: __1 mins__

## Changes proposed
Tell processor to allow the record even if attachments are missing

## Context for reviewers

https://www.elastic.co/guide/en/elasticsearch/reference/current/foreach-processor.html

ignore_missing is defaulted to False. This means if the record has no
attachments, it just drops it and moves on. This fix makes it not drop
it, and just allow it.

## Additional information
Set `ENABLE_OPPORTUNITY_ATTACHMENT_PIPELINE=false` in the local.env

Run `make db-seed-local` and `make populate-search-opportunities`,
without this fix, nothing ends up in the search index.