[Issue #3553] Modify local s3 setup in factories to be easier to use (#3574)
## Summary
Fixes #3553

### Time to review: __10 mins__

## Changes proposed
Modified our attachment factory to create files on s3 automatically

## Context for reviewers
We had a lot of tests and local setup assume files would already be
pre-generated that the opportunity attachment table would point to.
Those often were wrong and broke.

This fixes things by instead making the factories create s3 files for
you. It does this by taking the path you gave it, and making a file at
that location. This does make assumptions that your s3 bucket exists,
but I piggybacked on the `enable_factory_create` pytest fixture we have
that already requires you call it to call any DB operation with create,
so that shouldn't be too much of an issue. I've set it up so the factory
by default now creates no attachments on an opportunity, but can be told
to add a few because the performance degradation on running tests was
pretty bad.

## Additional information
Running `make db-seed-local` works. You can see the files it created
with `aws --endpoint-url http://localhost:4566 s3 ls
local-mock-public-bucket --recursive`

The tests all work with just a few minor fixes.