[Unticketed] Fix scenario where opp attachment is null (#3578)
## Summary

### Time to review: __3 mins__

## Changes proposed
Handle case (by erroring) if an opportunity attachment is null from
legacy data

## Context for reviewers
Note that by erroring like this, processing on other attachments will
continue.

Hit this issue when testing in dev, forgot that the `file_lob` column
that contains the opportunity attachments from the legacy data is a
nullable field. Easy enough fix to throw a value error so processing
continues and we skip that record.