[Issue #3540] AWS Pinpoint Mock setup (#3652)
## Summary
Fixes #3540 

### Time to review: __5 mins__

## Changes proposed
Adds basic setup to call AWS Pinpoint

Handles mocking queries for local development / unit tests

## Context for reviewers
AWS Pinpoint is an AWS service we'll be using for sending email
notifications to users. We don't yet have any of the infra setup for it,
but the Nava infra template does have it setup, and I used it to test
the behavior.

There will be more configuration needed in the future for things like
the app ID, but I wasn't sure exactly how we'll manage that (one app per
notification type, or just per env? Will leave to design to sort out).

We also won't send raw HTML emails and will instead setup email
templates most likely which will require adjusting some of the
parameters we pass in.

## Additional information
With a quick local script, I was able to pass in a few parameters and
send myself some emails:

![Screenshot 2025-01-27 at 3 39
36 PM](https://github.com/user-attachments/assets/750e9492-7560-4551-9d65-e7d37bba1b83)

I used this to verify the format of the response could be parsed / was
setup correctly.