docs: small updates to docs to optimize dev setup (#3790)
## Summary
Fixes N/A

### Time to review: 2 mins

## Changes proposed
As a new contributor, I noticed during my frontend dev setup that I
missed the "step" for copying over the `.env.development` to
`.env.local`. I believe this happened due to a small typo in the
original documentation (`/frontend/env.development` instead of
`/frontend/.env.development` - notice the **dot** in front of env). So I
just created the actual step of copying the environment variable file in
the numbered list of steps.

## Context for reviewers
I wanted to get some thoughts about this change. I'm hoping it can
optimize the speed of the local environment setup for new contributors.
Thank you!

## Additional information
N/A