[Issue #3780] Snapshot and disable writes (#3866)
## Summary
Fixes #3780

### Time to review: __15 mins__

## Changes proposed
Adding the syrupy plugin to the pytest framework to enable snapshot
testing for the ETL pipeline. We are testing the input of each step of
the top level ETL function up to the database. The snapshots

## Context for reviewers
> Testing instructions, background context, more in-depth details of the
implementation, and anything else you'd like to call out or ask
reviewers. Explain how the changes were verified.

The pipeline was ran with breakpoints to get sample data that went into
the input directory for the snapshots to compare to. The snapshots were
then generated using the syrupy framework and stored in the
__snapshots__ directory.

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.

<img width="719" alt="image"
src="https://github.com/user-attachments/assets/d2af819c-cb29-4122-96fd-009e03c64365"
/>

---------

Co-authored-by: David Dudas <david.dudas@intuitial.com>